fix(web): 일정 페이지에 서브 탭 도입 (캘린더/백로그/카테고리)
웹뷰에서 뒤로가기 버튼이 없어 별도 페이지로 이동하면 돌아올 수 없는 문제 해결.
백로그/카테고리를 별도 라우트 대신 일정 페이지 내 탭으로 통합하고,
기존 /backlog, /categories 경로는 /schedules로 리다이렉트 처리.
