+++ web/src/components/ui/app-shell.tsx
@@ -19,7 +19,7 @@ interface NavItem {
 }
 
 const NAV_ITEMS: NavItem[] = [
-  { href: '/schedules', label: '일정', Icon: CalendarIcon, activeOn: ['/backlog', '/categories'] },
+  { href: '/schedules', label: '일정', Icon: CalendarIcon },
   { href: '/routines', label: '루틴', Icon: ArrowPathIcon },
   { href: '/budget', label: '지출', Icon: WalletIcon },
 ];
